requirements.txt
mysql-connector-python>=8.1.0 